pyproject.toml
[project]
name = "django-lms"
version = "0.1.0"
description = ""
authors = [
    {name = "AntonSkyrda",email = "mr.skyrda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
